refactor: move theme provider to root
