+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import type { Viewport } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import type { ReactNode } from 'react'
 import CustomSearchDialog from '@/components/features/search'
-import { ThemeProvider } from '@/components/providers/theme'
+import { ThemeProvider } from '@/components/theme-provider'
 import { baseUrl } from '@/constants'
 import { socials } from '@/constants/navigation'
 import { description as homeDescription, owner, title } from '@/constants/site'
